pyproject.toml
[project]
name = "adkaiagentbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "databases>=0.9.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "bcrypt>=4.1.2",
    "tweepy>=4.15.0",
    "python-jose>=3.4.0",
    "openai>=1.75.0",
    "pinecone>=6.0.2",
    "tensorflow>=2.19.0",
    "openai-agents>=0.0.11",
]

[tool.setuptools]
packages = ["db", "models", "routes"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
psycopg2-binary = "^2.9.1"
python-dotenv = "^0.19.0"
tweepy = "^4.14.0"
